+++ src/logic/HttpError.js
@@ -1,0 +1,9 @@
+export class HttpError extends Error {
+  constructor(status, message) {
+    super(`HttpError`);
+    this.status = status;
+    this.message = message || '';
+    // Tiny hack to make extending error works
+    Object.setPrototypeOf(this, new.target.prototype);
+  }
+}
